Notifications: rework how heartbeats work.
They're now based on thread monitoring, so the "key" Objects in the API are now
implicitly the calling thread, and a heartbeat failure is automatically detected
if the thread exits without calling stopThreadHeartbeats(). Threads are also
responsible for periodically calling sendThreadHeartbeat(), so we can detect
deadlocks/hangs in addition to unexpected thread deaths.