Branding: implement config file uploads.
This adds a getOwnerName() method to the BrandPlugin abstract class, so
we can say "upload to <name>" instead of "upload to Micro-Manager.org".
It also massively cleans up the upload logic we used to have, as there's
no need to ensure a probably-unique filename on the client side (!). Let
the server handle that.